+++ src/commands/moderation/clear.shaii.ts
@@ -7,7 +7,7 @@ export default defineCommand({
   permissions: ["MANAGE_MESSAGES"],
   requiresProcessing: false,
   execute: async (message) => {
-    let count = parseInt(message.args[0]) + 1;
+    let count = parseFloat(message.args[0]) + 1;
     count = count > 100 ? 100 : count;
     // Will return if count is not a string.
     if (isNaN(count)) return `⚠️ when not number`;
